refactor: rename auth to user
